refactor(core): split model-resolver into SRP modules
Extract parseModelId/formatModelId into model-id.ts,
ProviderRegistry into provider-registry.ts,
ProviderHealthTracker into health-tracker.ts,
and createLLMStack into llm-stack.ts.
ModelResolver now has a single responsibility.
