+++ packages/core/src/llm/index.ts
@@ -1,10 +1,14 @@
 export * from './types.js';
-export * from './providers/anthropic.js';
-export * from './providers/openai.js';
-export * from './providers/ollama.js';
-export * from './providers/gemini.js';
+export * from './model-id.js';
+export * from './provider-registry.js';
+export * from './health-tracker.js';
 export * from './model-resolver.js';
+export * from './llm-stack.js';
 export * from './model-catalog.js';
 export * from './thinking-level.js';
 export * from './token-tracker.js';
 export * from './failover.js';
+export * from './providers/anthropic.js';
+export * from './providers/openai.js';
+export * from './providers/ollama.js';
+export * from './providers/gemini.js';
